requirements.txt
llama-cpp-python>=0.2.0
numpy>=1.24.0
